requirements.txt
pandas
numpy
requests
websocket-client
python-dotenv
google-cloud-secret-manager
